refactor: client type file imports
