+++ client/src/components/containers/CodeReviewMetricsContainers/CodeReviewMetricsGraphs/MetricsTrendGraphs/LeastVotesGraph/leastVotesGraphUtils.tsx
@@ -1,5 +1,6 @@
+import { IFetchedCodeReviewPullRequest } from "src/services/api/api";
+
 import { PullRequestsVotesAnalysis } from "./leastVotesGraphInterface";
-import { IFetchedCodeReviewPullRequest } from "../../../../../../fetchers/fetchers";
 import { Vote } from "../../../CodeReviewMetricsTable/interfaces";
 import {
   MAX_PULL_REQUEST_IDS_IN_LINE_IN_TOOLTIP,
