+++ client/src/components/containers/CodeReviewMetricsContainers/CodeReviewMetricsGraphs/MetricsTrendGraphs/TimeMetricsGraph/TimeMetricsGraph.tsx
@@ -1,11 +1,12 @@
 import { useState } from "react";
 
+import { IFetchedCodeReviewPullRequest } from "src/services/api/api";
+
 import { Monthly } from "./monthlyGraphUtils";
 import { GRAPH_COLOR, GRAPH_PLOT_NAME, GRAPH_TOOLTIP_HEADER, TREND_VIEW_CLASS } from "./timeMetricsConstants";
 import styles from "./TimeMetricsGraph.module.scss";
 import { IPullRequestsTimeMetrics, TimeMetrics } from "./timeMetricsGraphInterface";
 import { Weekly } from "./weeklyGraphUtils";
-import { IFetchedCodeReviewPullRequest } from "../../../../../../fetchers/fetchers";
 import { BarChart } from "../../../../../reusables/MetricsGraphs/BarChart/BarChart";
 import { GRAPH_TYPE, TREND_VIEW } from "../../../../../reusables/MetricsGraphs/BarChart/barChartConstants";
 import {
