+++ client/src/components/containers/CodeReviewMetricsContainers/CodeReviewMetricsGraphs/MetricsTrendGraphs/TimeMetricsGraph/timeMetricsGraphUtils.tsx
@@ -1,7 +1,7 @@
 import { SECONDS_IN_ONE_HOUR } from "src/constants/constants";
+import { IFetchedCodeReviewPullRequest } from "src/services/api/api";
 
 import { IPullRequestsTimeMetrics } from "./timeMetricsGraphInterface";
-import { IFetchedCodeReviewPullRequest } from "../../../../../../fetchers/fetchers";
 import { formatHoursToDays } from "../../../CodeReviewMetricsTiles/codeReviewMetricsTilesUtils";
 import {
   MAX_PULL_REQUEST_IDS_IN_LINE_IN_TOOLTIP,
