+++ client/src/components/containers/CodeReviewMetricsContainers/CodeReviewMetricsTable/CodeReviewMetricsTable.tsx
@@ -1,5 +1,7 @@
 import { useState, memo } from "react";
 
+import { IFetchedCodeReviewPullRequest } from "src/services/api/api";
+
 import styles from "./CodeReviewMetricsTable.module.scss";
 import { columns, sortMap } from "./codeReviewMetricsTableConstants";
 import { getFilteredPullRequests, getTotalComments, sortPullRequests } from "./codeReviewMetricsTableUtils";
@@ -7,7 +9,6 @@ import { CustomTableCell } from "./CustomTableCell";
 import { Vote, IColumn, Filters, FilterColumn } from "./interfaces";
 import { CodeReviewMetricsTableRow } from "./TableChildren/CodeReviewMetricsTableRow";
 import { VOTES_FILTER_DEFAULT_STATE } from "./VotesFilter/votesFilterConstants";
-import { IFetchedCodeReviewPullRequest } from "../../../../fetchers/fetchers";
 interface Props {
   pullRequests: IFetchedCodeReviewPullRequest[];
 }
