+++ client/src/context/clientFilters/clientFiltersContextTypes.ts
@@ -1,4 +1,4 @@
-import { IFetchedClientFilterSquadMember } from "../../fetchers/fetchers";
+import { IFetchedClientFilterSquadMember } from "src/services/api/api";
 
 export type IContextClientFilterSquadMembers = IFetchedClientFilterSquadMember[];
 export interface IContextClientFilterSquad {
