+++ client/src/pages/CodeReviewMetrics/CodeReviewMetrics.tsx
@@ -3,9 +3,9 @@ import { ChangeEvent, useEffect, useState } from "react";
 import { DateTime } from "luxon";
 import "react-datepicker/dist/react-datepicker.css";
 
-import { LoadingSpinner, SnackBar, TabToggle } from "src/components/components.ts";
-
+import { LoadingSpinner, SnackBar, TabToggle } from "src/components/components";
 import { useCodeReviewMetrics } from "src/services/api/api";
+
 import styles from "./CodeReviewMetrics.module.scss";
 import {
   CODE_REVIEW_METRICS,
